Full Release - The project is now open source.
Exported from the ground growing private branch master, which will be
depreciated in favor of this branch.

Changes include bugfixes and the addition of multithreaded realtime
updates.  The eventual goal is to be able to vary the texture resolution
and vertices based on need, storing textures and meshes if flagged for
later use.

Also I want to be able to render a flat plane of a planet based on
coordiates, and see if I can generate a more detailed map for use close
up (like on the surface of a planet).  As such the previous flat plane
renderer has been pulled out and I'm working on a new plane renderer to
take it's place.

Enjoy
-Richard